mcp-date-time/pyproject.toml
[project]
name = "datetime-mcp-server"
version = "0.1.0"
description = "MCP сервер для получения текущей даты и времени"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastmcp>=0.4.0",
    "uvicorn>=0.24.0",
    "starlette>=0.27.0",
    "pytz>=2023.3",
    "tzlocal>=5.0.1",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]